安慰小溯溯，爱门！

+++ Lib/MuRainLib.py
@@ -35,13 +35,22 @@ def reboot() -> None:
         sys.exit()
 
 
-def download_file_to_cache(url: str, headers=None) -> str | None:
+def download_file_to_cache(url: str, headers=None, file_name: str = "",
+                           download_path: str = None, stream=False) -> str | None:
     if headers is None:
         headers = {}
 
     # 路径拼接
-    file_name = url.split("/")[-1] + str(random.randint(10000, 99999)) + str(time.time()) + ".cache"
-    file_path = os.path.join(cache_path, file_name)
+    flag = False
+    if file_name == "":
+        file_name = url.split("/")[-1] + str(random.randint(10000, 99999)) + str(time.time()) + ".cache"
+    else:
+        flag = True
+
+    if download_path is None:
+        file_path = os.path.join(cache_path, file_name)
+    else:
+        file_path = os.path.join(download_path, file_name)
 
     # 路径不存在特判
     if not os.path.exists(cache_path):
@@ -49,32 +58,42 @@ def download_file_to_cache(url: str, headers=None) -> str | None:
 
     try:
         # 下载
-        with open(file_path, "wb") as f, requests.get(url, stream=True, headers=headers) as res:
-            for chunk in res.iter_content(chunk_size=64 * 1024):
-                if not chunk:
-                    break
-                f.write(chunk)
+        if stream:
+            with open(file_path, "wb") as f, requests.get(url, stream=True, headers=headers) as res:
+                for chunk in res.iter_content(chunk_size=64 * 1024):
+                    if not chunk:
+                        break
+                    f.write(chunk)
+        else:
+            # 不使用流式传输
+            res = requests.get(url, headers=headers, verify=False)
+
+            with open(file_path, "wb") as f:
+                f.write(res.content)
     except requests.exceptions.RequestException as e:
         logging.warning(f"下载文件失败: {e}")
         if os.path.exists(file_path):
             os.remove(file_path)
         return None
 
-    # 计算MD5
-    md5_hash = hashlib.md5()
-    with open(file_path, "rb") as f:
-        for byte_block in iter(lambda: f.read(4096), b""):
-            md5_hash.update(byte_block)
-        rename = md5_hash.hexdigest() + ".cache"
-        rename_path = os.path.join(cache_path, rename)
+    if not flag:
+        # 计算MD5
+        md5_hash = hashlib.md5()
+        with open(file_path, "rb") as f:
+            for byte_block in iter(lambda: f.read(4096), b""):
+                md5_hash.update(byte_block)
+            rename = md5_hash.hexdigest() + ".cache"
+            rename_path = os.path.join(cache_path, rename)
 
-    # 重命名（MD5）
-    if os.path.exists(rename_path):
-        os.remove(rename_path)
+        # 重命名（MD5）
+        if os.path.exists(rename_path):
+            os.remove(rename_path)
 
-    os.rename(file_path, rename_path)
+        os.rename(file_path, rename_path)
 
-    return rename_path
+        return rename_path
+    else:
+        return file_path
 
 
 # 删除缓存文件
